bug fixing

+++ tf_slim/fine_tune_voc2007_test.py
@@ -13,6 +13,7 @@ from constants import LOG_DIR
 from datasets import voc, imagenet
 from preprocess import preprocess_voc_new
 from utils import assign_from_checkpoint_fn, montage_tf
+import numpy as np
 
 slim = tf.contrib.slim
 
@@ -92,7 +93,7 @@ with sess.as_default():
 
         # Define learning parameters
         num_train_steps = 80000
-        boundaries = [10000, 20000, 30000, 40000, 50000, 60000, 70000]
+        boundaries = np.int64([10000, 20000, 30000, 40000, 50000, 60000, 70000])
         values = [0.001, 0.0005, 0.00025, 0.000125, 0.0000625, 0.00003125, 0.000015625, 0.0000078125]
         learning_rate = tf.train.piecewise_constant(global_step, boundaries=boundaries, values=values)
